Extra homepage "sections" spacing
We don't want to add explicit <section>s to the markdown as it makes
working on it a nightmare (anything in those sections also has to
become markup).  But we also want "sections" to stand off each other
a little, so add extra spacing to anything but the first homepage h2.

Co-authored-by: Tekin Süleyman <tekin@tekin.co.uk>